+++ http-client/Network/HTTP/Client/Types.hs
@@ -635,7 +635,7 @@ data Request = Request
     --
     -- @since 0.7.13
 
-    , earlyHintHeaderReceived :: Header -> IO ()
+    , earlyHintHeadersReceived :: [Header] -> IO ()
     -- ^ Called every time an HTTP 103 Early Hints header is received from the server.
     --
     -- @since 0.7.16